+++ src/main/java/net/mi/erp/dao/ProjectDAO.java
@@ -5,6 +5,6 @@ import com.landawn.abacus.query.SQLBuilder;
 
 import net.mi.erp.entity.Project;
 
-public interface ProjectDAO extends UncheckedCrudDao<Project, Long, SQLBuilder.PSC, ProjectDAO> {
+public interface ProjectDAO extends UncheckedCrudDao<Project, Integer, SQLBuilder.PSC, ProjectDAO> {
 
 }
